remove unused functions and move testing function to test
