+++ src/zkevm_specs/encoding/__init__.py
@@ -1,5 +1,2 @@
-from .constants import *
-from .typing import *
 from .utils import *
-from .constants import *
 from .lookup import LookupTable
