--- src/zkevm_specs/encoding/constants.py
@@ -1,2 +1,0 @@
-# This value is determined by the curve we use, which is BN254
-FIELD_SIZE = 2**254
