+++ src/zkevm_specs/util/__init__.py
@@ -7,7 +7,6 @@ from .constraint_system import *
 from .hash import *
 from .param import *
 from .typing import *
-from .testing import memory_expansion, memory_word_size
 
 
 def rand_range(stop: Union[int, float] = 2**256) -> int:
