+++ tests/evm/test_extcodecopy.py
@@ -24,10 +24,8 @@ from zkevm_specs.util import (
     keccak256,
     EXTRA_GAS_COST_ACCOUNT_COLD_ACCESS,
     GAS_COST_COPY,
-    memory_word_size,
-    memory_expansion,
 )
-
+from common import memory_expansion, memory_word_size
 
 TESTING_DATA = (
     # empty code
